fix(cluster): check route_rules for cluster references on delete
Previously cluster deletion only checked ai_route_rules, but the
effective routing rules are now in route_rules (global/entity/api-key).

- Move RouteRulesManager to model/route_rules package
- Add ClusterDeleteChecker to RouteRulesManager
- Register it in ClusterManager delete checkers
- Remove incorrect ClusterDeleteChecker from model/iai_route

Fixes #43
