+++ design-docs/sys-design/details/路由规则管理.md
@@ -388,7 +388,8 @@ API-Key 从一个 Entity 切换到另一个 Entity 时：
 
 | 文件 | 说明 |
 |------|------|
-| `model/shared/route_rules.go` | AI 路由规则业务模型与 CRUD、校验 |
+| `model/route_rules/route_rules.go` | Global/Entity/API-Key 路由规则业务模型与 CRUD、校验 |
+| `model/shared/types.go` | `RouteRulesParam`、`RouteRulesStorager`、`AiRouteRuleParam` 等结构定义 |
 | `model/shared/types.go` | `RouteRulesParam`、`AiRouteRuleParam` 等结构定义 |
 | `storage/rdb/quota/route_rules.go` | `route_rules` 表 RDB 存储实现 |
 | `storage/rdb/internal/dao/table_route_rules.go` | `route_rules` 表 DAO |
