+++ design-docs/sys-design/存储层设计文档.md
@@ -270,7 +270,7 @@ type EntityTypeStorager interface {
     DeleteEntityType(ctx context.Context, filter *EntityTypeFilter) error
 }
 
-// model/shared/route_rules.go
+// model/shared/types.go
 type RouteRulesStorager interface {
     CreateRouteRules(ctx context.Context, ruleType string, owner *string, param *RouteRulesParam) (int64, error)
     FetchRouteRules(ctx context.Context, ruleType string, owner *string) (*RouteRulesParam, error)
