+++ design-docs/sys-design/总体设计文档.md
@@ -288,7 +288,8 @@ ai-gateway-api/
 | `model/iprotocol` | TLS 证书模型与业务逻辑 |
 | `model/imods` | mod-api-key、mod-body-process、AI 路由导出配置组装 |
 | `model/quota` | EntityType、Entity、QuotaPlan、QuotaBalance、RateLimitPolicy、BalanceSync、QuotaResetScheduler |
-| `model/shared` | 跨包共享类型与 `RouteRulesManager` |
+| `model/shared` | 跨包共享类型与接口（含 `RouteRulesStorager`） |
+| `model/route_rules` | `RouteRulesManager`：Global/Entity/API-Key 路由规则业务逻辑 |
 | `model/itxn` | 事务抽象接口 `TxnStorager` |
 | `model/iversion_control` | 配置版本控制与增量导出 |
 
@@ -429,7 +430,7 @@ var (
     EntityManager *quota.EntityManager
     QuotaPlanManager *quota.QuotaPlanManager
     RateLimitPolicyManager *quota.RateLimitPolicyManager
-    RouteRulesManager *shared.RouteRulesManager
+    RouteRulesManager *route_rules.RouteRulesManager
     AIRouteExporter *imods.AIRouteExporter
     BalanceSyncManager *quota.BalanceSyncManager
     QuotaResetScheduler *quota.QuotaResetScheduler
