+++ design-docs/sys-design/模型层设计文档.md
@@ -1115,7 +1115,7 @@ type EntityManager struct {
     entityTypeMgr *EntityTypeManager
     quotaPlanMgr *QuotaPlanManager
     rateLimitPolicyMgr *RateLimitPolicyManager
-    routeRulesMgr *shared.RouteRulesManager
+    routeRulesMgr *route_rules.RouteRulesManager
 }
 
 func (m *EntityManager) CreateEntity(ctx context.Context, param *EntityParam) error {
@@ -1400,9 +1400,9 @@ func (s *QuotaResetScheduler) Start() {
 
 ---
 
-### 4.10 `model/shared`
+### 4.10 `model/shared` 与 `model/route_rules`
 
-跨包共享类型与 Storager 接口，供 API Key、Entity、Global 路由规则等共用。
+`model/shared` 提供跨包共享类型与 Storager 接口；`model/route_rules` 提供 Global/Entity/API-Key 路由规则业务逻辑。两者供 API Key、Entity、Global 路由规则等共用。
 
 ```go
 // model/shared/types.go
@@ -1446,9 +1446,9 @@ type BalanceSummary struct {
 ```
 
 ```go
-// model/shared/route_rules.go
+// model/shared/types.go
 const (
-    RouteRulesTypeAPIKey = "api_key"
+    RouteRulesTypeAPIKey = "apikey"
     RouteRulesTypeEntity = "entity"
     RouteRulesTypeGlobal = "global"
 )
@@ -1502,6 +1502,10 @@ type RouteRulesFilter struct {
     SortOrder *string `form:"sort_order"`
 }
 
+#### `model/route_rules/route_rules.go`
+
+```go
+// model/route_rules/route_rules.go
 type RouteRulesManager struct {
     txn itxn.TxnStorager
     storager RouteRulesStorager
