+++ endpoints/openapi_v1/global_route_rules/endpoints_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/yf-networks/ai-gateway-api/endpoints/openapi_v1/internal/testutil"
 	"github.com/yf-networks/ai-gateway-api/lib"
+	"github.com/yf-networks/ai-gateway-api/model/route_rules"
 	"github.com/yf-networks/ai-gateway-api/model/shared"
 	"github.com/yf-networks/ai-gateway-api/stateful/container"
 )
@@ -74,7 +75,7 @@ func (f *fakeRouteRulesStorager) FetchRouteRulesByID(ctx context.Context, id int
 
 func setupRouteRulesManager(storager shared.RouteRulesStorager) func() {
 	old := container.RouteRulesManager
-	container.RouteRulesManager = shared.NewRouteRulesManager(&testutil.FakeTxn{}, storager)
+	container.RouteRulesManager = route_rules.NewRouteRulesManager(&testutil.FakeTxn{}, storager)
 	return func() {
 		container.RouteRulesManager = old
 	}
